Rename config source package
