+++ pkg/lib/config/configsource/errors.go
@@ -1,4 +1,4 @@
-package source
+package configsource
 
 import "errors"
