+++ pkg/lib/config/configsource/local_fs.go
@@ -1,4 +1,4 @@
-package source
+package configsource
 
 import (
 	"context"
