+++ pkg/lib/config/configsource/source.go
@@ -1,4 +1,4 @@
-package source
+package configsource
 
 import (
 	"context"
